src/requirements.txt
alembic==1.0.3
arrow==0.13.0
atomicwrites==1.2.1
attrs==18.2.0
awscli==1.16.135
blis==0.2.4
boto3==1.9.86
botocore==1.12.125
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.8.1
colorama==0.3.9
config==0.4.0
coverage==4.5.3
cymem==2.0.2
dask==1.1.5
decorator==4.4.0
docutils==0.14
dominate==2.3.4
flake8==3.6.0
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
mkl-fft==1.0.10
mkl-random==1.0.2
more-itertools==5.0.0
murmurhash==1.0.2
networkx==2.2
numpy==1.16.2
nvidia-ml-py3==7.352.0
olefile==0.46
packaging==19.0
pandas==0.24.2
Pillow==5.4.1
plac==0.9.6
pluggy==0.9.0
preshed==2.0.1
py==1.7.0
pyasn1==0.4.5
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
pyparsing==2.3.1
pytest==4.4.1
pytest-cov==2.6.1
pytest-flask==0.14.0
pytest-parallel==0.0.9
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
pytz==2018.9
PyWavelets==1.0.2
PyYAML==3.13
requests==2.21.0
rsa==3.4.2
s3transfer==0.2.0
scikit-image==0.14.2
scipy==1.2.1
selenium==3.141.0
six==1.12.0
spacy==2.1.3
SQLAlchemy==1.2.14
srsly==0.0.5
thinc==7.0.4
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tqdm==4.31.1
typing==3.6.6
urllib3==1.24.1
visitor==0.1.3
wasabi==0.2.1
Werkzeug==0.14.1
wincertstore==0.2
WTForms==2.2.1
